Create Chapter 1 Day 1
Explain what the Blockchain is in your own words.

The Blockchain is a shared decentralized database that store data in blocks and is linked through cryptography.  It is safe and secure and is the forefront of the new tech age.  


Explain what a Smart Contract is.

Smart contracts are programs stored on the blockchain that have certain functions that are performed when conditions within the contract are met.  


Explain the difference between a script and a transaction.

A scripts is a read only to view data.  Whereas a transaction is a pay call.