Create Chapter 1 Day 2
What are the 5 Cadence Programming Language Pillars?

1 Safety and Security / 2 Clarity / 3 Approachability  / 4 Developer Experience / 5 Resource Oriented Programming 


In your opinion, even without knowing anything about the Blockchain or coding, why could the 5 Pillars be useful?

The 5 Pillars are useful due to the need for each of these.  You need it to be safe and secure for it to be trusted.  The clarity and the approachability  for coding is most important since that is the language.  The error messages have already helped me in my quest for chapter 2.